ref: rebuild server

+++ src/broker.c
@@ -8,46 +8,76 @@
 #include "lib/socket.h"
 #include "lib/queue.h"
 
-#define FILEPATH "msg.sock"
-#define BUFFER_SIZE 64
-#define MAX_CLIENTS 4
+#define FILEPATH "sock.msg"
 
-int file_exists(const char* file_path) {
+bool file_exists(const char* file_path) {
     struct stat file_stat;
     return (stat(file_path, &file_stat) == 0);
 }
 
 void handle_sigint(int sig) {
-    printf("\ncleaning up...\n");
+    if (!file_exists(FILEPATH)) exit(EXIT_FAILURE);
+    printf("unlinking %s and exiting.", FILEPATH);
     unlink(FILEPATH);
     exit(EXIT_SUCCESS);
 }
 
-void start_server(char* file_path, int buffer_size) {
+void register_connection(int socket_fd, struct pollfd* clients, int* num_clients, int max_clients) {
+    if (clients[0].revents & POLLIN) {
+        int client_socket_fd = accept_connection(socket_fd);
+        if (*num_clients < max_clients) {
+            clients[(*num_clients) + 1].fd = client_socket_fd;
+            clients[(*num_clients) + 1].events = POLLIN;
+            (*num_clients)++;
+            printf("new client connected - total clients: %d\n", *num_clients);
+        } else {
+            printf("party is full\n");
+            close(client_socket_fd);
+        }
+    }
+}
+
+void redirect_messages(struct pollfd* clients, int* num_clients, char* buffer, int buffer_size) {
+    for (int i = 1; i <= *num_clients; i++) {
+        if (clients[i].revents && POLLIN) {
+            int num_bytes = receive_message(clients[i].fd, buffer, buffer_size);
+            if (num_bytes <= 0) {
+                // client disconnect or error
+                close(clients[i].fd);
+                printf("client disconnected.\n");
+                clients[i] = clients[*num_clients];
+                (*num_clients)--;
+                i--;
+            } else {
+                // forward the message to all other clients
+                buffer[num_bytes] = '\0';
+                for (int j = 1; j <= *num_clients; j++) {
+                    if (j != i)
+                        send_message(clients[j].fd, buffer, num_bytes);
+                }
+            }
+        }
+    }
+}
+
+void start_server(char* filepath, int buffer_size, int max_clients) {
     char buffer[buffer_size];
-    
-    int server_socket_fd;
-    int client_socket_fd;
+    int socket_fd;
     int num_clients = 0;
-    struct sockaddr_un address;
-    struct pollfd clients[MAX_CLIENTS];
+    struct pollfd clients[max_clients];
 
-    if (file_exists(file_path)) unlink(file_path);
+    if (file_exists(filepath)) unlink(filepath);
 
-    create_socket(&server_socket_fd, &address, file_path);
-    bind_socket(server_socket_fd, &address);
-    listen_socket(server_socket_fd);
+    socket_fd = create_server(filepath);
 
-    printf("server bound to %s\nwaiting for connections...\n", file_path);
+    printf("waiting for connections.\n");
 
-    signal(SIGINT, handle_sigint);
-
-    clients[0].fd = server_socket_fd;
+    clients[0].fd = socket_fd;
     clients[0].events = POLLIN;
 
-    for (int i = 1; i < MAX_CLIENTS; i++) {
-        clients[i].fd = -1;
-    }
+    for (int i = 1; i < max_clients; i++) clients[i].fd = -1;
+
+    signal(SIGINT, handle_sigint);
 
     // main loop to handle requests
     while (true) {
@@ -59,45 +89,15 @@ void start_server(char* file_path, int buffer_size) {
         }
 
         // check if there is a new connection
-        if (clients[0].revents & POLLIN) {
-            client_socket_fd = accept_connection(server_socket_fd);
-            if (num_clients < MAX_CLIENTS) {
-                clients[num_clients + 1].fd = client_socket_fd;
-                clients[num_clients + 1].events = POLLIN;
-                num_clients++;
-                printf("new client connected. Total clients: %d\n", num_clients);
-            } else {
-                printf("party is full\n");
-                close(client_socket_fd);
-            }
-        }
+        register_connection(socket_fd, clients, &num_clients, max_clients);
 
         // check each client for incoming messages
-        for (int i = 1; i <= num_clients; i++) {
-            if (clients[i].revents && POLLIN) {
-                int num_bytes = receive_message(clients[i].fd, buffer, BUFFER_SIZE);
-                if (num_bytes <= 0) {
-                    // client disconnect or error
-                    close(clients[i].fd);
-                    printf("client disconnected.\n");
-                    clients[i] = clients[num_clients];
-                    num_clients--;
-                    i--;
-                } else {
-                    // forward the message to all other clients
-                    buffer[num_bytes] = '\0';
-                    for (int j = 1; j <= num_clients; j++) {
-                        if (j != i)
-                            send_message(clients[j].fd, buffer, num_bytes);
-                    }
-                }
-            }
-        }
+        redirect_messages(clients, &num_clients, buffer, buffer_size);
     }
 }
 
 int main(int argc, char* argv[]) {
-    start_server(FILEPATH, BUFFER_SIZE);
+    start_server(FILEPATH, 128, 4);
 
     return 0;
 }
